Single implementation to create FilterToken vector
- indices_of does not work for multiple occurences of the same flag.

+++ cli/src/verbosity.rs
@@ -1,10 +1,10 @@
-static mut VERBOSITY_LEVEL: u32 = 0;
+static mut VERBOSITY_LEVEL: u8 = 0;
 
 pub fn verbosity() -> bool {
     let v = unsafe { VERBOSITY_LEVEL };
     v > 0
 }
 
-pub fn set_verbosity(v: u64) {
-    unsafe { VERBOSITY_LEVEL = v as u32; }
+pub fn set_verbosity(v: u8) {
+    unsafe { VERBOSITY_LEVEL = v; }
 }
